Refactor `getCalculatedMeasuresFromSavedWidgets`
